feat: move and rename project name and update docs
