+++ Masa.Contrib.sln
@@ -27,6 +27,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{42DF7AAC-362
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "solution items", "solution items", "{9F6F9899-D5F1-444A-BE56-64F949550D22}"
 	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
 		nuget.config = nuget.config
 	EndProjectSection
 EndProject
@@ -172,7 +173,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Storage", "Storage", "{1653
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.Contrib.Configuration.AutoMap.NoArgumentConstructor.Tests", "test\Masa.Contrib.Configuration.AutoMap.NoArgumentConstructor.Tests\Masa.Contrib.Configuration.AutoMap.NoArgumentConstructor.Tests.csproj", "{B8358ED1-C95A-4EC0-9756-FB32C931F204}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.Contrib.Observability.Opentelemetry", "src\Observability\Masa.Contrib.Observability.Opentelemetry.Logging\Masa.Contrib.Observability.Opentelemetry.csproj", "{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.Contrib.BasicAbility.Tsc", "src\BasicAbility\Masa.Contrib.BasicAbility.Tsc\Masa.Contrib.BasicAbility.Tsc.csproj", "{EB0F6BDA-C6AA-4859-834E-5613BB38862B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -678,14 +679,14 @@ Global
 		{B8358ED1-C95A-4EC0-9756-FB32C931F204}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B8358ED1-C95A-4EC0-9756-FB32C931F204}.Release|x64.ActiveCfg = Release|Any CPU
 		{B8358ED1-C95A-4EC0-9756-FB32C931F204}.Release|x64.Build.0 = Release|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Debug|x64.Build.0 = Debug|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Release|x64.ActiveCfg = Release|Any CPU
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A}.Release|x64.Build.0 = Release|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Debug|x64.Build.0 = Debug|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Release|x64.ActiveCfg = Release|Any CPU
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -771,7 +772,7 @@ Global
 		{97532A33-A591-4DF5-A2C0-72527B78ED82} = {38E6C400-90C0-493E-9266-C1602E229F1B}
 		{165391A5-034E-4894-8084-8DF7D4AA7518} = {42DF7AAC-362C-48F4-B76A-BDEEEFF17CC9}
 		{B8358ED1-C95A-4EC0-9756-FB32C931F204} = {9EEE31DA-3165-4CB3-AAE9-27CC3A4DE669}
-		{51B8F2EC-EC53-43B1-A016-F4A84E9F0B5A} = {75050CBC-A0F2-408A-A582-54EF37450B29}
+		{EB0F6BDA-C6AA-4859-834E-5613BB38862B} = {5DFAF4A2-ECB5-46E4-904D-1EA5F48B2D48}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {40383055-CC50-4600-AD9A-53C14F620D03}
