+++ src/BasicAbility/Masa.Contrib.BasicAbility.Tsc/Logging/MasaLoggingOptions.cs
@@ -1,3 +1,6 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
 namespace Masa.Contrib.Observability;
 
 public class MasaLoggingOptions
