+++ src/BasicAbility/Masa.Contrib.BasicAbility.Tsc/Logging/ResourceBuilderExtenstions.cs
@@ -1,3 +1,6 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
 namespace OpenTelemetry.Resources;
 
 public static class ResourceBuilderExtenstions
@@ -12,7 +15,7 @@ public static class ResourceBuilderExtenstions
 
         if (!string.IsNullOrEmpty(options.ProjectName))
         {
-            resourceBuilder.AddAttributes(new KeyValuePair<string, object>[] { new KeyValuePair<string, object>(MasaResourceSemanticConventions.AttributeServiceProjectName, options.ProjectName) });
+            resourceBuilder.AddAttributes(new KeyValuePair<string, object>[] { new KeyValuePair<string, object>(OpenTelemetryAttributeName.Service.ProjectName, options.ProjectName) });
         }
         return resourceBuilder;
     }
