+++ src/BasicAbility/Masa.Contrib.BasicAbility.Tsc/Tracing/ActivityExtensions.cs
@@ -1,3 +1,6 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
 namespace System.Diagnostics;
 
 public static class ActivityExtension
