+++ src/BasicAbility/Masa.Contrib.BasicAbility.Tsc/Tracing/AspNetCoreInstrumentationOptionsExtensions.cs
@@ -1,3 +1,6 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
 namespace Microsoft.Extensions.DependencyInjection;
 
 public static class AspNetCoreInstrumentationOptionsExtensions
