+++ src/BasicAbility/Masa.Contrib.BasicAbility.Tsc/_Imports.cs
@@ -1,14 +1,16 @@
+// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
+global using Masa.Contrib.BasicAbility.Tsc;
 global using Masa.Contrib.Observability;
 global using Microsoft.AspNetCore.Http;
 global using OpenTelemetry.Contrib.Instrumentation.ElasticsearchClient;
 global using OpenTelemetry.Contrib.Instrumentation.EntityFrameworkCore;
-global using OpenTelemetry.Exporter;
 global using OpenTelemetry.Instrumentation.AspNetCore;
 global using OpenTelemetry.Instrumentation.Http;
 global using OpenTelemetry.Instrumentation.StackExchangeRedis;
 global using OpenTelemetry.Logs;
 global using OpenTelemetry.Metrics;
-global using OpenTelemetry.Resources;
 global using OpenTelemetry.Trace;
 global using StackExchange.Redis;
 global using System.Diagnostics;
