--- src/Observability/Masa.Contrib.Observability.Opentelemetry.Logging/Logging/MasaResourceSemanticConventions.cs
@@ -1,6 +1,0 @@
-namespace OpenTelemetry.Resources;
-
-internal class MasaResourceSemanticConventions
-{
-    public const string AttributeServiceProjectName = "service.project.name";
-}
