--- src/Observability/Masa.Contrib.Observability.Opentelemetry.Logging/Metrics/MasaServiceExtensions.cs
@@ -1,21 +1,0 @@
-namespace Microsoft.Extensions.DependencyInjection;
-
-public static partial class MasaServiceExtensions
-{
-    public static IServiceCollection AddMasaMetrics(this IServiceCollection services, Action<MeterProviderBuilder> builderConfiger)
-    {
-        services.AddOpenTelemetryMetrics(builder =>
-           {
-               builder
-               .AddRuntimeMetrics()
-               .AddAspNetCoreInstrumentation()
-               .AddHttpClientInstrumentation();
-
-               if (builderConfiger != null)
-                   builderConfiger.Invoke(builder);
-           }
-        );
-
-        return services;
-    }
-}
